requirements.txt
requests~=2.28.1
PyMySQL~=1.0.2
Flask~=2.2.2
opencv-python~=4.6.0.66
numpy~=1.23.4
torch~=1.13.0
tqdm~=4.64.1
sklearn~=0.0.post1
scikit-learn~=1.1.3
matplotlib~=3.6.2
Pillow~=9.3.0
scikit-image~=0.19.3
librosa~=0.9.2
scipy~=1.9.3
pydub~=0.25.1
torchvision~=0.14.0
cpca~=0.5.5